tunniplaan, täiendused

+++ hois_back/src/main/java/ee/hitsa/ois/web/BuildingController.java
@@ -37,53 +37,56 @@ public class BuildingController {
     // building: get/save/update/delete
     @GetMapping("/buildings/{id:\\d+}")
     public BuildingDto getBuilding(HoisUserDetails user, @WithEntity("id") Building building) {
-        UserUtil.assertSameSchool(user, building.getSchool());
+        UserUtil.assertIsSchoolAdmin(user, building.getSchool());
         return BuildingDto.of(building);
     }
 
     @PostMapping("/buildings")
     public HttpUtil.CreatedResponse createBuilding(HoisUserDetails user, @Valid @RequestBody BuildingForm form) {
+        UserUtil.assertIsSchoolAdmin(user);
         return HttpUtil.created(buildingService.create(user, form));
     }
 
     @PutMapping("/buildings/{id:\\d+}")
     public BuildingDto updateBuilding(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestBody = true) Building building, @Valid @RequestBody BuildingForm form) {
-        UserUtil.assertSameSchool(user, building.getSchool());
+        UserUtil.assertIsSchoolAdmin(user, building.getSchool());
         return getBuilding(user, buildingService.save(building, form));
     }
 
     @DeleteMapping("/buildings/{id:\\d+}")
     public void deleteBuilding(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestParam = "version") Building building, @SuppressWarnings("unused") @RequestParam("version") Long version) {
-        UserUtil.assertSameSchool(user, building.getSchool());
+        UserUtil.assertIsSchoolAdmin(user, building.getSchool());
         buildingService.delete(building);
     }
 
     // room: search/get/save/update/delete
     @GetMapping("/rooms")
     public Page<RoomSearchDto> searchRooms(HoisUserDetails user, @Valid RoomSearchCommand searchCommand, Pageable pageable) {
+        UserUtil.assertIsSchoolAdmin(user);
         return buildingService.findAllRooms(user.getSchoolId(), searchCommand, pageable);
     }
 
     @GetMapping("/rooms/{id:\\d+}")
     public RoomDto getRoom(HoisUserDetails user, @WithEntity("id") Room room) {
-        UserUtil.assertSameSchool(user, room.getBuilding().getSchool());
+        UserUtil.assertIsSchoolAdmin(user, room.getBuilding().getSchool());
         return RoomDto.of(room);
     }
 
     @PostMapping("/rooms")
     public HttpUtil.CreatedResponse createRoom(HoisUserDetails user, @Valid @RequestBody RoomForm form) {
+        UserUtil.assertIsSchoolAdmin(user);
         return HttpUtil.created(buildingService.create(user, form));
     }
 
     @PutMapping("/rooms/{id:\\d+}")
     public RoomDto updateRoom(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestBody = true) Room room, @Valid @RequestBody RoomForm form) {
-        UserUtil.assertSameSchool(user, room.getBuilding().getSchool());
+        UserUtil.assertIsSchoolAdmin(user, room.getBuilding().getSchool());
         return getRoom(user, buildingService.save(user, room, form));
     }
 
     @DeleteMapping("/rooms/{id:\\d+}")
     public void deleteRoom(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestParam = "version") Room room, @SuppressWarnings("unused") @RequestParam("version") Long version) {
-        UserUtil.assertSameSchool(user, room.getBuilding().getSchool());
+        UserUtil.assertIsSchoolAdmin(user, room.getBuilding().getSchool());
         buildingService.delete(room);
     }
 }
